Refactor AccordionItemBody (and wrapper) to tsx
